Kontrolor.

+++ WebApp/Models/Vozilo.cs
@@ -9,8 +9,8 @@ namespace WebApp.Models
     {
         public int Id { get; set; }
         public string Tip { get; set; }
-        public int X { get; set; }
-        public int Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
 
         public int LinijaId { get; set; }
         public Linija Linija { get; set; }
